Create Make Sum of Array Zero
Chef has an array A of length N consisting of 1 and −1 only. In one operation, Chef can choose any index i such that 1 ≤ i ≤N and multiply the element Ai​ by −1. Find the minimum number of operations required to make the sum of the array equal to 0. Return-1 if the sum of the array cannot be made 0.